Give ball's velocity attributes more appropriate names
It's more comment practice to refer to an objects velocity along a given
axis "vn" where "n" is the axis (x or y in this case). Originally was
thinking of these attribute names in terms of "max velocity", hence the
"m" in the original names... but that was a dumb name.