Optimistically move ball; update related code accordingly
Also this just moves us closer to having better variable names in this
confusing function. Also, fixed two possible collision bugs where we
were checking if overlap < 0 instead of > 0 and where we weren't
accounting for, in the case of a side bounce, that vx had been reversed
(so we have to re-reverse it to calculate hitx in that case).